Set up league_stats_helper_test
