+++ lib/modules/league_stat_helper_module.rb
@@ -190,18 +190,4 @@ module LeagueStatHelpers
     end
     difference_home_vs_away_won
   end
-
-  def most_goals_scored(id)
-    team_games = @game_teams.select do |game|
-       game.team_id == id
-     end
-    team_games.max_by {|game| game.goals}.goals
-  end
-
-  def fewest_goals_scored(id)
-    team_games = @game_teams.select do |game|
-       game.team_id == id
-     end
-    team_games.min_by {|game| game.goals}.goals
-  end
 end
